optimise hydration

+++ template/hydration_test.go
@@ -685,7 +685,7 @@ func TestMapToDict(t *testing.T) {
 			},
 			expected: []map[string]any{
 				{"key": "string"},
-				{"key": float64(123)}, // Using float64 since JSON unmarshalling converts numbers to float64
+				{"key": 123}, // No longer converts to float64 since we removed JSON round-trip
 				{"key": true},
 				{"key": nil},
 			},
@@ -1921,7 +1921,7 @@ func TestTemplateWithAddkeytoallFunction(t *testing.T) {
 			expected: []any{
 				map[string]any{"ID": "1", "content": "memory 1", "resource_id": "resource-123"},
 				"not a dict",
-				float64(123), // Use float64 since JSON unmarshaling converts numbers to float64
+				123, // No longer converts to float64
 			},
 		},
 	}
