+++ template/map_to_array_test.go
@@ -56,7 +56,7 @@ func TestMapToArray(t *testing.T) {
 			},
 			expected: []map[string]any{
 				{"key": "bool_key", "value": true},
-				{"key": "int_key", "value": float64(123)},
+				{"key": "int_key", "value": 123},
 				{"key": "nil_key", "value": nil},
 				{"key": "string_key", "value": "string_value"},
 			},
